requirements.txt
yfinance
pandas
numpy
CurrencyConverter
PyYAML
seaborn
matplotlib
networkx
PyPortfolioOpt
quantstats